UseLogin added
Added a new property that controls whether or not we send the site_login
element and its contents at all. The API is very picky about the values
contained so if you're not going to use the values then you have to omit
the entire section.

I changed the task to either send that segment or not based on the value
of a new property UseLogin. I updated the readme to reflect this.